requirements.txt
# Core Data Science & ML (Priority #1)
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
prophet>=1.1.5
statsmodels>=0.14.0
scipy>=1.10.0
joblib>=1.3.0

# AWS / Cloud Integration (Priority #3)
boto3>=1.34.0
sagemaker>=2.200.0
pyarrow>=14.0.0          # For efficient Parquet I/O with Athena/S3
moto>=4.2.0              # For mocking AWS services in tests

# Visualization & Reporting (Capacity Planning visuals)
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0
dash>=2.14.0
reportlab>=4.0.0         # PDF reports
openpyxl>=3.1.0          # Excel exports

# Utilities & Production Support
pyyaml>=6.0.0            # Config loading
python-dotenv>=1.0.0     # .env file handling
argparse                 # Built-in, but listed for clarity
pytest>=7.4.0            # Unit testing
logging                  # Built-in
requests>=2.31.0         # Any API calls if needed

# Jupyter Lab + Notebook Ecosystem
jupyterlab>=4.0.0
notebook>=7.0.0
ipywidgets>=8.1.0        # Interactive widgets in notebooks
jupyter_contrib_nbextensions>=0.7.0  # Optional extra notebook extensions